test: migrate tests for login and participant components, and expense datasource
