+++ app/mikane/src/app/pages/login/login.component.spec.ts
@@ -1,4 +1,5 @@
-import { ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
+import { Component } from '@angular/core';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { MatButtonModule } from '@angular/material/button';
 import { MatCardModule } from '@angular/material/card';
 import { MatFormFieldModule } from '@angular/material/form-field';
@@ -8,80 +9,56 @@ import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { By } from '@angular/platform-browser';
 import { Router, provideRouter } from '@angular/router';
-import { RouterTestingHarness } from '@angular/router/testing';
-import { MockComponent, MockModule, MockService } from 'ng-mocks';
 import { of, throwError } from 'rxjs';
-import { spyPropertyGetter } from 'src/app/helpers/test.helpers';
+import { beforeEach, describe, expect, it, vi } from 'vitest';
 import { AuthService } from 'src/app/services/auth/auth.service';
 import { BreakpointService } from 'src/app/services/breakpoint/breakpoint.service';
 import { LogService } from 'src/app/services/log/log.service';
 import { MessageService } from 'src/app/services/message/message.service';
 import { User } from 'src/app/services/user/user.service';
 import { ApiError } from 'src/app/types/apiError.type';
-import { AccountComponent } from '../account/account.component';
-import { EventsComponent } from '../events/events.component';
 import { LoginComponent } from './login.component';
+
+@Component({ selector: 'app-events', template: '' })
+class MockEventsComponent {}
+
+@Component({ selector: 'app-account', template: '' })
+class MockAccountComponent {}
 
 describe('Login Component', () => {
 	let component: LoginComponent;
 	let fixture: ComponentFixture<LoginComponent>;
 	let page: PageObject;
-	let authService: jasmine.SpyObj<AuthService>;
-	let messageService: jasmine.SpyObj<MessageService>;
-	let breakpointService: jasmine.SpyObj<BreakpointService>;
+
+	let authService: { login: ReturnType<typeof vi.fn>, logout: ReturnType<typeof vi.fn>, sendResetPasswordEmail: ReturnType<typeof vi.fn>, getCurrentUser: ReturnType<typeof vi.fn>, redirectUrl: string | null };
+	let messageService: { showSuccess: ReturnType<typeof vi.fn>, showError: ReturnType<typeof vi.fn> };
+	let breakpointService: { isMobile: ReturnType<typeof vi.fn> };
 
 	beforeEach(async () => {
-		authService = jasmine.createSpyObj<AuthService>(
-			'AuthService',
-			['login', 'logout', 'sendResetPasswordEmail', 'getCurrentUser'],
-			['redirectUrl'],
-		);
-		messageService = jasmine.createSpyObj<MessageService>('MessageService', ['showSuccess', 'showError']);
-		breakpointService = jasmine.createSpyObj<BreakpointService>('BreakpointService', ['isMobile']);
+		authService = { login: vi.fn(), logout: vi.fn(), sendResetPasswordEmail: vi.fn(), getCurrentUser: vi.fn(), redirectUrl: null };
+		messageService = { showSuccess: vi.fn(), showError: vi.fn() };
+		breakpointService = { isMobile: vi.fn() };
 
 		TestBed.configureTestingModule({
-			imports: [LoginComponent],
+			imports: [LoginComponent, MatFormFieldModule, MatInputModule, MatProgressSpinnerModule, MatButtonModule, MatToolbarModule, MatCardModule, MatIconModule],
 			providers: [
 				{ provide: AuthService, useValue: authService },
 				{ provide: MessageService, useValue: messageService },
 				{ provide: BreakpointService, useValue: breakpointService },
-				{ provide: LogService, useValue: MockService(LogService) },
+				{ provide: LogService, useValue: { log: vi.fn(), error: vi.fn() } },
 				provideRouter([
 					{ path: 'login', loadComponent: () => import('./login.component').then((m) => m.LoginComponent) },
-					{ path: 'events', component: MockComponent(EventsComponent) },
-					{ path: 'account', component: MockComponent(AccountComponent) },
+					{ path: 'events', component: MockEventsComponent },
+					{ path: 'account', component: MockAccountComponent },
 				]),
 			],
-		}).overrideComponent(LoginComponent, {
-			remove: {
-				imports: [
-					MatFormFieldModule,
-					MatInputModule,
-					MatProgressSpinnerModule,
-					MatButtonModule,
-					MatToolbarModule,
-					MatCardModule,
-					MatIconModule,
-				],
-			},
-			add: {
-				imports: [
-					MockModule(MatInputModule),
-					MockModule(MatProgressSpinnerModule),
-					MockModule(MatButtonModule),
-					MockModule(MatToolbarModule),
-					MockModule(MatCardModule),
-					MockModule(MatIconModule),
-				],
-			},
-		});
-		await RouterTestingHarness.create('/login');
+		}).compileComponents();
 
 		fixture = TestBed.createComponent(LoginComponent);
-		page = new PageObject(fixture);
 		component = fixture.componentInstance;
+		page = new PageObject(fixture);
 
-		(fixture.debugElement.injector.get(BreakpointService) as jasmine.SpyObj<BreakpointService>).isMobile.and.returnValue(of(false));
+		breakpointService.isMobile.mockReturnValue(of(false));
 		fixture.detectChanges();
 	});
 
@@ -96,8 +73,7 @@ describe('Login Component', () => {
 	});
 
 	it('should login user', () => {
-		authService.login.and.returnValue(of({} as User));
-		messageService.showSuccess.and.stub();
+		authService.login.mockReturnValue(of({} as User));
 
 		page.usernameInput.value = 'username';
 		page.passwordInput.value = 'password';
@@ -107,18 +83,17 @@ describe('Login Component', () => {
 		page.loginButton.nativeElement.click();
 		fixture.detectChanges();
 
-		expect(fixture.debugElement.injector.get(AuthService).login).toHaveBeenCalledOnceWith('username', 'password');
-		expect(fixture.debugElement.injector.get(MessageService).showSuccess).toHaveBeenCalledOnceWith('Login successful');
-		expect(fixture.debugElement.injector.get(MessageService).showError).not.toHaveBeenCalled();
+		expect(authService.login).toHaveBeenCalledWith('username', 'password');
+		expect(messageService.showSuccess).toHaveBeenCalledWith('Login successful');
+		expect(messageService.showError).not.toHaveBeenCalled();
 	});
 
 	it('should fail to login user if login info is wrong', () => {
-		authService.login.and.returnValue(
+		authService.login.mockReturnValue(
 			throwError(() => {
 				return { error: { code: 'PUD-003' } } as ApiError;
 			}),
 		);
-		messageService.showError.and.stub();
 
 		page.usernameInput.value = 'wrong username';
 		page.passwordInput.value = 'wrong password';
@@ -127,14 +102,13 @@ describe('Login Component', () => {
 		page.loginButton.nativeElement.click();
 		fixture.detectChanges();
 
-		expect(fixture.debugElement.injector.get(AuthService).login).toHaveBeenCalledOnceWith('wrong username', 'wrong password');
-		expect(fixture.debugElement.injector.get(MessageService).showError).toHaveBeenCalledOnceWith('Wrong username or password');
-		expect(fixture.debugElement.injector.get(MessageService).showSuccess).not.toHaveBeenCalled();
+		expect(authService.login).toHaveBeenCalledWith('wrong username', 'wrong password');
+		expect(messageService.showError).toHaveBeenCalledWith('Wrong username or password');
+		expect(messageService.showSuccess).not.toHaveBeenCalled();
 	});
 
 	it('should show error if no user returned', () => {
-		authService.login.and.returnValue(of(undefined));
-		messageService.showError.and.stub();
+		authService.login.mockReturnValue(of(undefined));
 
 		page.usernameInput.value = 'username';
 		page.passwordInput.value = 'password';
@@ -144,7 +118,7 @@ describe('Login Component', () => {
 		page.loginButton.nativeElement.click();
 		fixture.detectChanges();
 
-		expect(messageService.showError).toHaveBeenCalledOnceWith('Login failed');
+		expect(messageService.showError).toHaveBeenCalledWith('Login failed');
 	});
 
 	it('should focus on username field if fields empty', () => {
@@ -155,10 +129,9 @@ describe('Login Component', () => {
 		expect(page.focus.nativeElement).toBe(page.usernameInput);
 	});
 
-	it('should redirect to events page after login', fakeAsync(() => {
-		authService.login.and.returnValue(of({} as User));
-		authService.getCurrentUser.and.returnValue(of({} as User));
-		messageService.showSuccess.and.stub();
+	it('should redirect to events page after login', async () => {
+		authService.login.mockReturnValue(of({} as User));
+		authService.getCurrentUser.mockReturnValue(of({} as User));
 
 		page.usernameInput.value = 'username';
 		page.passwordInput.value = 'password';
@@ -166,15 +139,17 @@ describe('Login Component', () => {
 		page.passwordInput.dispatchEvent(new Event('input'));
 
 		page.loginButton.nativeElement.click();
-		tick();
-		fixture.detectChanges();
+		await fixture.whenStable();
 
 		expect(TestBed.inject(Router).url).toEqual('/events');
-	}));
+	});
 
-	it('should redirect to stored redirect url after login', fakeAsync(() => {
-		authService.login.and.returnValue(of({} as User));
-		spyPropertyGetter(authService, 'redirectUrl').and.returnValue('/account');
+	it('should redirect to stored redirect url after login', async () => {
+		authService.login.mockReturnValue(of({} as User));
+		Object.defineProperty(authService, 'redirectUrl', {
+			get: () => '/account',
+			set: vi.fn(),
+		});
 
 		page.usernameInput.value = 'username';
 		page.passwordInput.value = 'password';
@@ -182,23 +157,22 @@ describe('Login Component', () => {
 		page.passwordInput.dispatchEvent(new Event('input'));
 
 		page.loginButton.nativeElement.click();
-		tick();
-		fixture.detectChanges();
+		await fixture.whenStable();
 
 		expect(TestBed.inject(Router).url).toEqual('/account');
-	}));
+	});
 
 	it('should toggle forgot password view', () => {
 		expect(page.loginForm).toBeDefined();
 		expect(page.resetForm).toBeFalsy();
-		expect(page.forgotPasswordButton.innerText).toBe('Forgot password?');
+		expect(page.forgotPasswordButton.textContent).toBe('Forgot password?');
 
 		page.forgotPasswordButton.click();
 		fixture.detectChanges();
 
 		expect(page.resetForm).toBeDefined();
 		expect(page.loginForm).toBeFalsy();
-		expect(page.forgotPasswordButton.innerText).toBe('Back');
+		expect(page.forgotPasswordButton.textContent).toBe('Back');
 
 		page.forgotPasswordButton.click();
 		fixture.detectChanges();
@@ -208,7 +182,7 @@ describe('Login Component', () => {
 	});
 
 	it('should request password reset', () => {
-		authService.sendResetPasswordEmail.and.returnValue(of(undefined));
+		authService.sendResetPasswordEmail.mockReturnValue(of(undefined));
 		page.forgotPasswordButton.click();
 		fixture.detectChanges();
 
@@ -220,12 +194,12 @@ describe('Login Component', () => {
 		page.passwordButton.click();
 		fixture.detectChanges();
 
-		expect(authService.sendResetPasswordEmail).toHaveBeenCalledOnceWith('email@email.com');
+		expect(authService.sendResetPasswordEmail).toHaveBeenCalledWith('email@email.com');
 		expect(page.infoText).toBeDefined();
 	});
 
 	it('should show error response if request fails', () => {
-		authService.sendResetPasswordEmail.and.returnValue(
+		authService.sendResetPasswordEmail.mockReturnValue(
 			throwError(() => {
 				return {
 					status: 400,
@@ -245,9 +219,9 @@ describe('Login Component', () => {
 		fixture.detectChanges();
 
 		expect(page.errorText).toBeDefined();
-		expect(page.errorText.nativeElement.innerHTML).toContain('Server not configured for sending email');
+		expect(page.errorText.nativeElement.textContent).toContain('Server not configured for sending email');
 
-		authService.sendResetPasswordEmail.and.returnValue(
+		authService.sendResetPasswordEmail.mockReturnValue(
 			throwError(() => {
 				return {
 					status: 500,
@@ -262,11 +236,11 @@ describe('Login Component', () => {
 		fixture.detectChanges();
 
 		expect(page.errorText).toBeDefined();
-		expect(page.errorText.nativeElement.innerText).toBe('Something went wrong while sending email :(');
+		expect(page.errorText.nativeElement.textContent.trim()).toBe('Something went wrong while sending email :(');
 	});
 
 	it('should reset error text when toggling forgot password view', () => {
-		authService.sendResetPasswordEmail.and.returnValue(
+		authService.sendResetPasswordEmail.mockReturnValue(
 			throwError(() => {
 				return {
 					status: 400,
@@ -295,7 +269,7 @@ describe('Login Component', () => {
 
 	describe('on mobile', () => {
 		beforeEach(() => {
-			(fixture.debugElement.injector.get(BreakpointService) as jasmine.SpyObj<BreakpointService>).isMobile.and.returnValue(of(true));
+			breakpointService.isMobile.mockReturnValue(of(true));
 			fixture.detectChanges();
 		});
